requirements.txt
# mgd 2025.03.06
# python>=3.13.0
beautifulsoup4==4.13.3
email-validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Minify==0.49
Flask-SQLAlchemy==3.1.1
Flask_WTF==1.2.2
google-auth>=2.30.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.140.0
Jinja2==3.1.6
pipreqs>=0.4.13
psycopg2-binary==2.9.10
requests==2.32.3
sendgrid==6.11.0
SQLAlchemy==2.0.38
# type checking in VS Code/Pylance, install types-Flask.
types-Flask==3.0.3.*
Werkzeug==3.1.3
WTForms==3.2.1
puremagic==1.28.0
#eof